Rearrange validator function
